Extract forum work request route contract
